Fix cross platform loading of textures.
Currently it's not possible to load textures from a material file that
was exported on a different platform than the one it's being loaded on.
This change uses the path of the XML file as the base bath for relative
textures instead of reading the platform specific path stored in the
file.